+++ apiv2/test_infra/factories/annotation.py
@@ -17,6 +17,7 @@ from faker_enum import EnumProvider
 from platformics.test_infra.factories.base import CommonFactory
 from test_infra.factories.deposition import DepositionFactory
 from test_infra.factories.run import RunFactory
+from test_infra.factories.tomogram_voxel_spacing import TomogramVoxelSpacingFactory
 
 Faker.add_provider(Bioseq)
 Faker.add_provider(Organ)
@@ -35,6 +36,9 @@ class AnnotationFactory(CommonFactory):
     )
     deposition = factory.SubFactory(
         DepositionFactory,
+    )
+    tomogram_voxel_spacing = factory.SubFactory(
+        TomogramVoxelSpacingFactory,
     )
     s3_metadata_path = fuzzy.FuzzyText()
     https_metadata_path = fuzzy.FuzzyText()