+++ apiv2/test_infra/factories/annotation_shape.py
@@ -33,7 +33,16 @@ class AnnotationShapeFactory(CommonFactory):
         AnnotationFactory,
     )
     shape_type = fuzzy.FuzzyChoice(
-        ["SegmentationMask", "OrientedPoint", "Point", "InstanceSegmentation", "InstanceSegmentationMask", "Mesh"],
+        [
+            "SegmentationMask",
+            "OrientedPoint",
+            "Point",
+            "InstanceSegmentation",
+            "InstanceSegmentationMask",
+            "Mesh",
+            "GlobalCaption",
+            "AnnotationCaption",
+        ],
     )
 
     # Auto increment integer identifiers starting with 1